build: Added build option -Dcachedir with default value '/var/cache'
This is to replace a hard coded path to the cache directory. Ideally put
into configuration file. Since this is the cache directory of the binary
configuration file, creates a chicken and egg situation. End up having
to pass the path somehow. Might as well be at build vs runtime option.