daemon: Remove odd code to end session when logged in
Remove code to end a session before starting entrance when logged in.
Which is the only time the ENTRACE_USER env variable is set. This seems
to be a very odd scenario, not likely encountered. Not sure what good
ending the session does. When it will try to start another entrance
process or something abnormal. The variable would not be set in other
instances of entrance outside the users session/env.

Can restore if needed. Seems this is a code path that will never be
followed, and thus serves no purpose.