daemon: Switch to system killall client from kill client pid, invalid pid
The handle returned and resulting pid are of the shell process not the
actual client process. Which causes only termination of the shell
process and orphans client. This is not ideal but means to kill the
client on server termination, and no orphan client. Shell terminates
itself after client, and normal shutdown occurs.

Ideally return back to kill via pid, either looked up from /proc or
sending from client in an event. For now taking easier approach till
decided on further action. This maybe suffice unless a system lacks
killall.