bin: Terimate client on failure to connect to server initial fixes for #17
Refactored function to return boolean value of client server connection.
If client fails to connect return false and shutdown client. Which in
turn terminates entrance entirely. That is more desirable to represent a
problem than to give a false impression entrance is working. Both client
and server started, but not connected. Causing client not function
properly with elements missing.

Initial fixes for #17
https://github.com/Obsidian-StudiosInc/entrance/issues/17