bin: Nest code in conditional, for when init fails but shutdown is called
Present code does not make sense as it uses _gui before a null check.
Which the null check could be redundant and removed. Howevr it maybe
possible to call shutdown even when init fails. Thus a safer approach is
to nest all usage of _gui in a null check conditional. That ensures at
least _gui is not null for any usage of members of that struct pointer,
not just when free'd.