bin: Remove unused incomplete unusable initial vkbd support
This seems to be based on an old virtual keyboard that status is not
known. It has not been developed in sometime. Nothing is using the code
in this file. Thus pretty sure it was never implemented, added to build
system for compiling and linking.

Virtual keyboard library this was based on, but was incomplete.
https://git.enlightenment.org/libs/ekbd.git